fix: extract shared RoleBadge component (#92)
Move RoleBadge and ROLE_STYLES to shared component.
Remove duplicate implementations from UsersPanel and ApiKeyDisplay.
Addresses Qodo rule violation from PR #90 review.
